setup app/server/events directory; api/create-participant emits an event, server/events/notify-of-new-participants acts on it
